reuse filter in home page
